+++ packages/moonstone/ContextualPopupDecorator/ContextualPopupDecorator.js
@@ -35,7 +35,7 @@ const defaultConfig = {
 	/**
 	 * Disables passing the `skin` prop to the wrapped component.
 	 *
-	 * @see ui/Skinnable.Skinnable.skin
+	 * @see moonstone/Skinnable.Skinnable.skin
 	 * @type {Boolean}
 	 * @default false
 	 * @memberof moonstone/ContextualPopupDecorator.ContextualPopupDecorator.defaultConfig
@@ -199,7 +199,7 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * [ContextualPopup]{@link moonstone/ContextualPopupDecorator.ContextualPopup} and not
 			 * to the popup's activator component.
 			 *
-			 * @see ui/Skinnable.Skinnable.skin
+			 * @see moonstone/Skinnable.Skinnable.skin
 			 * @type {String}
 			 * @public
 			 */