+++ packages/moonstone/internal/Picker/PickerButton.js
@@ -8,7 +8,6 @@ import Touchable from '@enact/ui/Touchable';
 
 import Icon from '../../Icon';
 import IconButton from '../../IconButton';
-import {withSkinnableProps} from '../../Skinnable';
 
 import css from './Picker.less';
 
@@ -44,7 +43,6 @@ const PickerButtonBase = kind({
 		]),
 		joined: PropTypes.bool,
 		onSpotlightDisappear: PropTypes.func,
-		skin: PropTypes.string,
 		spotlightDisabled: PropTypes.bool
 	},
 
@@ -83,7 +81,6 @@ const PickerButtonBase = kind({
 		if (joined) {
 			delete rest.hidden;
 			delete rest.onSpotlightDisappear;
-			delete rest.skin;
 			delete rest.spotlightDisabled;
 
 			return (
@@ -100,9 +97,7 @@ const PickerButtonBase = kind({
 });
 
 const PickerButton = Pure(
-	withSkinnableProps(
-		PickerButtonBase
-	)
+	PickerButtonBase
 );
 
 export default PickerButton;